+++ Source/Krypton Components/Krypton.Toolkit/Translations/Converters/PaletteModeStrings.cs
@@ -99,6 +99,10 @@ public class PaletteModeStrings : GlobalId
     internal const string DEFAULT_PALETTE_MATERIAL_LIME_GREEN_DARK = @"Material - Lime Green - Dark Mode";
     internal const string DEFAULT_PALETTE_MATERIAL_LIME_GREEN_RIPPLE = @"Material - Lime Green (Ripple)";
     internal const string DEFAULT_PALETTE_MATERIAL_LIME_GREEN_DARK_RIPPLE = @"Material - Lime Green - Dark Mode (Ripple)";
+    internal const string DEFAULT_PALETTE_MATERIAL_DARK_GRAY = @"Material - Dark Gray";
+    internal const string DEFAULT_PALETTE_MATERIAL_LIGHT_GRAY = @"Material - Light Gray";
+    internal const string DEFAULT_PALETTE_MATERIAL_DARK_GRAY_RIPPLE = @"Material - Dark Gray (Ripple)";
+    internal const string DEFAULT_PALETTE_MATERIAL_LIGHT_GRAY_RIPPLE = @"Material - Light Gray (Ripple)";
     internal const string DEFAULT_PALETTE_MATERIAL_HIGH_CONTRAST = @"Material - High Contrast";
     internal const string DEFAULT_PALETTE_MATERIAL_DEUTERANOPIA = @"Material - Deuteranopia";
     internal const string DEFAULT_PALETTE_MATERIAL_PROTANOPIA = @"Material - Protanopia";
@@ -162,8 +166,8 @@ public class PaletteModeStrings : GlobalId
             { DEFAULT_PALETTE_OFFICE_2007_PROTANOPIA, PaletteMode.Office2007Protanopia },
             { DEFAULT_PALETTE_OFFICE_2007_LIME_GREEN, PaletteMode.Office2007LimeGreen },
             { DEFAULT_PALETTE_OFFICE_2007_LIME_GREEN_DARK, PaletteMode.Office2007LimeGreenDark },
-            //{ DEFAULT_PALETTE_OFFICE_2007_DARK_GRAY, PaletteMode.Office2007DarkGray },
-            //{ PaletteModeStrings.DEFAULT_PALETTE_OFFICE_2007_LIGHT_GRAY, PaletteMode.Office2007LightGray },
+            { DEFAULT_PALETTE_OFFICE_2007_DARK_GRAY, PaletteMode.Office2007DarkGray },
+            { DEFAULT_PALETTE_OFFICE_2007_LIGHT_GRAY, PaletteMode.Office2007LightGray },
             { DEFAULT_PALETTE_OFFICE_2010_BLUE, PaletteMode.Office2010Blue },
             { DEFAULT_PALETTE_OFFICE_2010_BLUE_DARK_MODE, PaletteMode.Office2010BlueDarkMode },
             { DEFAULT_PALETTE_OFFICE_2010_BLUE_LIGHT_MODE, PaletteMode.Office2010BlueLightMode },
@@ -178,8 +182,8 @@ public class PaletteModeStrings : GlobalId
             { DEFAULT_PALETTE_OFFICE_2010_PROTANOPIA, PaletteMode.Office2010Protanopia },
             { DEFAULT_PALETTE_OFFICE_2010_LIME_GREEN, PaletteMode.Office2010LimeGreen },
             { DEFAULT_PALETTE_OFFICE_2010_LIME_GREEN_DARK, PaletteMode.Office2010LimeGreenDark },
-            //{ DEFAULT_PALETTE_OFFICE_2010_DARK_GRAY, PaletteMode.Office2010DarkGray },
-            //{ PaletteModeStrings.DEFAULT_PALETTE_OFFICE_2010_LIGHT_GRAY, PaletteMode.Office2010LightGray },
+            { DEFAULT_PALETTE_OFFICE_2010_DARK_GRAY, PaletteMode.Office2010DarkGray },
+            { DEFAULT_PALETTE_OFFICE_2010_LIGHT_GRAY, PaletteMode.Office2010LightGray },
             { DEFAULT_PALETTE_OFFICE_2013_DARK_GRAY, PaletteMode.Office2013DarkGray },
             { DEFAULT_PALETTE_OFFICE_2013_LIGHT_GRAY, PaletteMode.Office2013LightGray },
             { DEFAULT_PALETTE_OFFICE_2013_WHITE, PaletteMode.Office2013White },
@@ -210,8 +214,8 @@ public class PaletteModeStrings : GlobalId
             { DEFAULT_PALETTE_MICROSOFT_365_BLACK_DARK_MODE_ALTERNATE, PaletteMode.Microsoft365BlackDarkModeAlternate},
             { DEFAULT_PALETTE_MICROSOFT_365_LIME_GREEN, PaletteMode.Microsoft365LimeGreen },
             { DEFAULT_PALETTE_MICROSOFT_365_LIME_GREEN_DARK, PaletteMode.Microsoft365LimeGreenDark },
-            //{ DEFAULT_PALETTE_MICROSOFT_365_DARK_GRAY, PaletteMode.Microsoft365DarkGray },
-            //{ PaletteModeStrings.DEFAULT_PALETTE_MICROSOFT_365_LIGHT_GRAY, PaletteMode.Microsoft365LightGray },
+            { DEFAULT_PALETTE_MICROSOFT_365_DARK_GRAY, PaletteMode.Microsoft365DarkGray },
+            { DEFAULT_PALETTE_MICROSOFT_365_LIGHT_GRAY, PaletteMode.Microsoft365LightGray },
             { DEFAULT_PALETTE_VISUAL_STUDIO_2010_RENDER_2007, PaletteMode.VisualStudio2010Render2007 },
             { DEFAULT_PALETTE_VISUAL_STUDIO_2010_RENDER_2010, PaletteMode.VisualStudio2010Render2010 },
             { DEFAULT_PALETTE_VISUAL_STUDIO_2010_RENDER_2013, PaletteMode.VisualStudio2010Render2013 },
@@ -244,6 +248,10 @@ public class PaletteModeStrings : GlobalId
             { DEFAULT_PALETTE_MATERIAL_LIME_GREEN_DARK, PaletteMode.MaterialLimeGreenDark },
             { DEFAULT_PALETTE_MATERIAL_LIME_GREEN_RIPPLE, PaletteMode.MaterialLimeGreenRipple },
             { DEFAULT_PALETTE_MATERIAL_LIME_GREEN_DARK_RIPPLE, PaletteMode.MaterialLimeGreenDarkRipple },
+            { DEFAULT_PALETTE_MATERIAL_DARK_GRAY, PaletteMode.MaterialDarkGray },
+            { DEFAULT_PALETTE_MATERIAL_LIGHT_GRAY, PaletteMode.MaterialLightGray },
+            { DEFAULT_PALETTE_MATERIAL_DARK_GRAY_RIPPLE, PaletteMode.MaterialDarkGrayRipple },
+            { DEFAULT_PALETTE_MATERIAL_LIGHT_GRAY_RIPPLE, PaletteMode.MaterialLightGrayRipple },
             { DEFAULT_PALETTE_MATERIAL_HIGH_CONTRAST, PaletteMode.MaterialHighContrast },
             { DEFAULT_PALETTE_MATERIAL_DEUTERANOPIA, PaletteMode.MaterialDeuteranopia },
             { DEFAULT_PALETTE_MATERIAL_PROTANOPIA, PaletteMode.MaterialProtanopia },